Add SlashSyntax.unparse
There was previously no way to convert a ScopedKey into a string that
could be parsed by the Act.aggregatedKeyParser. Because slash syntax
requires a capitalized name (more generally there is no requirement that
the slash syntax identifier matches the config name), it was necessary
to add a method toConfigIdent to KeyIndex so that we can invert the
config name to find the identifier.