Use swoval source formatting plugin
This switches from using the combination of the scalameta sbt-scalafmt and
the swoval sbt-java-format plugins to just using the swoval
sbt-jvm-format plugin. The jvm format plugin has a pair of nice
features:

1. It provides a command jvmfmtAll (and jvmfmtCheckAll) which
effectively formats all of the java, scala and sbt source files in the
project, including those files that comprise the sbt build definition.
2. It integrates with supershell so that the user can see which files
are currently being formatted (the number of formatting tasks is
configurable but defaults to the number of available processors divided
by 2).

The api of the swoval scalafmt plugin is very similar to the scalameta
plugin. It also auto-updates by parsing the scalafmt version from the
scalafmt.conf file and uses coursier to fetch the jars needed to perform
formatting for the given version.