reduce BuildServerReporter verbosity
The BSP protocol introduce a 'reset' field in the publishDiagnostic notification
that makes the cumulation of errors at client side possible.

Consequently it is not needed to maintain the `problemByFiles` map in the reporter.

The reporter first reset the diagnostic of each file, then publish the problems
one at a time.